Build and Pack Configuration.

+++ build.bat
@@ -1,4 +1,5 @@
-del /F /Q dist\*
+if exist "dist" del /F /Q dist\*
+if not exist "dist" mkdir dist
 cd src\
 python setup.py py2exe > ../dist/build.log 2>&1
